refactor: packages/api-legacy added enum
