+++ packages/api-legacy/enum/reservation-status.ts
@@ -1,0 +1,6 @@
+export enum ReservationStatus { 
+    Pending = "pending", 
+    Active = "active", 
+    Cancelled = "cancelled" , 
+    Archived = "archived" 
+};
