+++ packages/api-legacy/models/listing.ts
@@ -1,15 +1,11 @@
 import { pgQuery } from "../config/pgclient";
-import { Listing } from "./listing.schema";
+import { Listing, ListingProps } from "../interface/models";
 
 export async function getListings({
   offset,
   limit,
   featured,
-}: {
-  offset: number;
-  limit: number;
-  featured: boolean;
-}): Promise<Listing[]> {
+}: Partial<ListingProps>): Promise<Listing[]> {
   const query = `SELECT * FROM listings WHERE is_featured = $1 ORDER BY id LIMIT $2 OFFSET $3`;
   const params = [featured, limit, offset];
 
@@ -17,7 +13,7 @@ export async function getListings({
   return listings.rows.map(listingMapper);
 }
 
-export async function getListingById({ id }: { id: string | undefined }): Promise<Listing> {
+export async function getListingById({ id }: Partial<ListingProps>): Promise<Listing> {
   const listing = await pgQuery(`SELECT * FROM listings WHERE id = $1 LIMIT 1`, [id]);
   return listing.rows.map(listingMapper)[0];
 }
