+++ packages/api-legacy/models/payment.ts
@@ -1,13 +1,13 @@
-import PaymentModel, { Payment } from "./payment.schema";
+import PaymentModel from "./payment.schema";
+import { Payment } from "../interface/models";
+
+type IPaymentStatus = "pending" | "declined" | "completed" | "cancelled";
 
 export async function savePayment(payment: Partial<Payment>): Promise<Payment> {
   return PaymentModel.create(payment);
 }
 
-export async function updatePaymentStatus(
-  id: string,
-  status: "pending" | "declined" | "completed" | "cancelled",
-): Promise<Payment | null> {
+export async function updatePaymentStatus(id: string, status: IPaymentStatus): Promise<Payment | null> {
   const record = await PaymentModel.findOne({ _id: id });
   if (record) {
     record.status = status;
