--- packages/api-legacy/models/reservation-request.ts
@@ -1,7 +1,0 @@
-export interface ReservationRequest {
-  userId: string;
-  listingId: string;
-  from: string;
-  to: string;
-  guests: number;
-}
