+++ packages/api-legacy/models/reservation.ts
@@ -1,18 +1,18 @@
-import ReservationModel, { Reservation } from "./reservation.schema";
+import ReservationModel from "./reservation.schema";
+import { Reservation } from "../interface/models";
+import { ReservationStatus } from "../enum/reservation-status";
 
 export async function saveReservation(reservation: Partial<Reservation>): Promise<Reservation> {
   return ReservationModel.create(reservation);
 }
 
-export async function updateReservationStatus(
-  id: string,
-  status: "pending" | "active" | "cancelled" | "archived",
-): Promise<Reservation | null> {
+export async function updateReservationStatus(id: string, status: ReservationStatus): Promise<Reservation | null> {
   const record = await ReservationModel.findOne({ _id: id });
   if (record) {
     if (
-      (status === "active" && record.status !== "pending") ||
-      (status === "cancelled" && record.status !== "pending")
+      (status === ReservationStatus.Active && record.status !== ReservationStatus.Pending) ||
+      (status === ReservationStatus.Cancelled && record.status !== ReservationStatus.Pending)
+    
     ) {
       throw new Error(`Invalid reservation status transition: ${record.status} -> ${status}`);
     }
