+++ packages/api/src/models/favorite.ts
@@ -1,8 +1,7 @@
 import { pgQuery } from "../config/pgclient";
 import FavoriteModel from "./favorite.schema";
-import { Favorite } from "../interface/models";
 import { listingMapper } from "./listing";
-import { Listing } from "../interface/models";
+import { Favorite, Listing } from "../interface/models";
 
 export async function saveFavorite(fav: Favorite): Promise<Favorite | null> {
   const recordFavorite = await findFavorite(fav);
